Add entropy coding subcommands for bb_sga.py
Specifically, this adds the subcommands `encode_latents`  and
`decode_latents` to `bb_sga.py`.

Unfortunately, encoding and decoding for the bitsback model has to be
done on CPU because GPU calculations turned out to be not exactly
reproducible between encoder and decoder. Luckily, only the rate
optimization (last step) is necessary for encoding and decoding, the
(more expensive) RD-optimization is part of the `compress` subcommand,
which can still run on GPU.